+++ R/check.class.R
@@ -1,6 +1,7 @@
 check.class <- function(obj, class, list.of.class=TRUE) {
   #' Function to check whether obj belongs to class or
   #' is a list of class (if list.of.class==TRUE)
+  #' @keywords internal
   
   err.message1 <- paste("Input must be a", class, "object")
   err.message2 <- paste(err.message1, "or a list of", class, "objects")